--- src/collection/auth/mod.rs
@@ -1,4 +1,0 @@
-mod auth_refresh;
-mod auth_with_password;
-mod impersonate;
-mod request_verification;
