+++ src/records/crud/get_full_list.rs
