+++ src/records/crud/mod.rs
@@ -1,0 +1,6 @@
+mod create;
+mod get_first_list_item;
+mod get_full_list;
+mod get_list;
+mod get_one;
+mod update;
