+++ src/records/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod auth;
+pub(crate) mod crud;
